# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FileBrowser Backend API",
    "description": "Full API reference and interactive testing interface for the FileBrowser Node.js/Express backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "XAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token (without 'Bearer ' prefix)."
      },
      "XAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth",
        "description": "Alternative authorization header using 'X-Auth'."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "scope": { "type": "string" },
          "locale": { "type": "string" },
          "viewMode": { "type": "string" },
          "singleClick": { "type": "boolean" },
          "lockPassword": { "type": "boolean" },
          "hideDotfiles": { "type": "boolean" },
          "dateFormat": { "type": "boolean" },
          "commands": {
            "type": "array",
            "items": { "type": "string" }
          },
          "rules": {
            "type": "array",
            "items": { "type": "string" }
          },
          "perm": {
            "type": "object",
            "properties": {
              "admin": { "type": "boolean" },
              "execute": { "type": "boolean" },
              "create": { "type": "boolean" },
              "rename": { "type": "boolean" },
              "modify": { "type": "boolean" },
              "delete": { "type": "boolean" },
              "share": { "type": "boolean" },
              "download": { "type": "boolean" }
            }
          },
          "createdAt": { "type": "integer" },
          "updatedAt": { "type": "integer" }
        }
      },
...
      "FileInfo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "size": { "type": "integer" },
          "path": { "type": "string" },
          "url": { "type": "string" },
          "isDir": { "type": "boolean" },
          "modified": { "type": "string" },
          "type": { "type": "string" },
          "mode": { "type": "integer" },
          "isSymlink": { "type": "boolean" },
          "extension": { "type": "string" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FileInfo" }
          }
        }
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate User",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "admin" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Username and password required" },
          "403": { "description": "Invalid credentials" }
        }
      }
    },
    "/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User created successfully" },
          "400": { "description": "Validation failed" },
          "405": { "description": "Signup is disabled" },
          "409": { "description": "Username already taken" }
        }
      }
    },
    "/renew": {
      "post": {
        "tags": ["Auth"],
        "summary": "Renew current JWT token",
        "responses": {
          "200": {
            "description": "Token renewed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/resources": {
      "get": {
        "tags": ["Resources"],
        "summary": "List files and directories in root",
        "responses": {
          "200": {
            "description": "List of files/directory structure",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileInfo" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Upload file or create directory in root",
        "parameters": [
          {
            "name": "override",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Overwrite file if it exists"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            },
            "application/octet-stream": {
              "schema": { "type": "string", "format": "binary" }
            }
          }
        },
        "responses": {
          "200": { "description": "Upload/Creation successful" },
          "403": { "description": "Permission denied" }
        }
      }
    },
    "/resources/{path}": {
      "get": {
        "tags": ["Resources"],
        "summary": "List sub-directory files or get sub-file info",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "File or folder path relative to user scope"
          },
          {
            "name": "checksum",
            "in": "query",
            "schema": { "type": "string", "enum": ["md5", "sha1", "sha256", "sha512"] },
            "description": "Calculate and return file checksum"
          },
          {
            "name": "x-encoding",
            "in": "header",
            "schema": { "type": "boolean" },
            "description": "Set to true to download text content raw"
          }
        ],
        "responses": {
          "200": {
            "description": "FileInfo returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileInfo" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Upload file or create directory at specific path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "File path, or folder path ending in '/'"
          },
          {
            "name": "override",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Overwrite existing file"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            },
            "application/octet-stream": {
              "schema": { "type": "string", "format": "binary" }
            }
          }
        },
        "responses": {
          "200": { "description": "File/Directory created successfully" },
          "409": { "description": "File already exists" }
        }
      },
      "put": {
        "tags": ["Resources"],
        "summary": "Overwrite/Update file at path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            },
            "application/octet-stream": {
              "schema": { "type": "string", "format": "binary" }
            }
          }
        },
        "responses": {
          "200": { "description": "File updated successfully" },
          "404": { "description": "File not found" }
        }
      },
      "delete": {
        "tags": ["Resources"],
        "summary": "Delete file or directory at path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No content, deleted successfully" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "tags": ["Resources"],
        "summary": "Copy or rename/move file/directory",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "destination",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Target path relative to scope"
          },
          {
            "name": "action",
            "in": "query",
            "schema": { "type": "string", "enum": ["rename", "copy"], "default": "rename" }
          },
          {
            "name": "override",
            "in": "query",
            "schema": { "type": "boolean" }
          },
          {
            "name": "rename",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Auto-suffix if destination exists"
          }
        ],
        "responses": {
          "200": { "description": "Successful operation" },
          "404": { "description": "Source not found" },
          "409": { "description": "Destination exists" }
        }
      }
    },
...
  }
}